+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - AbstractDataProvider: throw 500 instead of 400 in case a filter exception occurs on creation
   of a filter from a prepared filter (defined in bundle config).
+- Rest/Query/Filter: also allow filter shortcuts with only path and operator for operators that don't have/need a value:
+  "filter\[path\]\[operator\]=IS_NULL"
 
 ## v0.1.242